REVERT THIS COMMIT - Temporarily update dbt-adapters/dbt-tests-adapters dependencies
This is necessary because the changes we are making in dbt-adapters via https://github.com/dbt-labs/dbt-adapters/pull/331
only get tested when run with an adapter. So this branch on dbt-postgres is not just
for ensuring the test is run, but also proof for the open PR in dbt-adapters that
the tests work. To show that though, we have to update the dependencies to temporarily
point at the dbt-adapters branch that has the work (because the work hasn't been merged
to main yet)